Add ripser and giotto-tda Mapper to README
Thought that adding Ripser to the list of mentions for other PH packages would be a good idea.  I am also proposing to add a mention to the implementation of Mapper in the giotto-tda package, but I am a co-creator and maintainer of that package so there is a glaring conflict of interest in me doing so.